Update README.md
2020 is the year of the turn.
After a careful analysis of the market, after evaluating business strategies, consolidating partnerships and making the necessary changes to make SG20 Token a different tool, from the others, but still based on concrete and consolidated foundations of Blockchain and Cryptocurrencies, it passes the baton to its design and programming team, creating the Smart Contract. 
Ethereum and its technology ERC20, become the basis of the entire SGN ecosystem, ecosystem that allows not only the creation of the Token "SG20" but all that will be the development, increase and distribution of the same.

Ethereum, following a careful and scrupulous verification, allows SG20 to position itself on the Etherscan.io platform. The procedure duly registered in Blockchain allowed to start the pre-launch phase on November 9, 2020